Add comparison operators
The `expected` class until this point has not been comparable
with itself, or with the underlying value type. This API omission
results in consumers of this type needing to test
`x.has_value() && x.value() == y`, which is unnecessarily verbose.

To account for this, separate comparison operators have been added.