✨ Reimplement expected in C++11
The original `expected<T,U>` implementation was written in C++17,
however many environments are still not able to work in C++17, and
thus cannot make use of this utility.

Nothing about implementing `expected<T,E>` actually *requires* C++17;
it's just simple to implement with `std::variant`. However, to make
this utility more accessible to consumers in earlier C++ versions, this
has been completely rewritten to be compatible with C++11.