🔨 Restructure `expected_base`
The base class, `expected_base`, introduces a bit of a conflict with
`detail::extract_error`. `extract_error` works on `expected`, and
requires access to the private base `expected_base` to retrieve
the `m_error`. However, in order to access this from `expected`
requires upcasting to `expected_base` -- which can only be done
with `reinterpet_cast`, which cannot be done in a `constexpr` context.

To fix this, the base class has been changed to a member rather than
a base. This prevents the need for `reinterpret_cast`, and also
should fix the `constexpr` issues.