🩹 Fix C++11 dangling reference
GCC is detecting that ternary statements that are throwing objects are
producing rvalue references that are unable to bind to lvalue
references.

Prior to a CWG defect report, this behavior is actually accurate, since
for a ternary statement only returns an lvalue if all branches return
an lvalue -- and `throw`ing an exception results in the object returned
to be by-value. This results in a temporary object that gets bound to
a reference -- hence the diagnostic messages claiming that it is
returning a reference to a temporary.

In order to preserve the `constexpr` requirements of being a single
return statement in C++11, this has been implemented using
the or-operator and using the comma operator. This guarantees that
exceptions only throw if `has_value()` is `false`, and ternary is
not used resulting in a temporary.