🩹 Fix `constexpr` `void` functions in C++11
`constexpr` functions in C++11 require literal return types to be valid.
Some compilers are liberal and allow for this, other ones complain and
refuse to build.

This modifies the `void` functions in C++11 to no longer be `constexpr`.

The trivial implementation of `destroy()` was using `constexpr` while
returning a `void` type. Some versions of clang and GCC accept this,
however some versions also warning that in `-std=c++11`, this is invalid
for `constexpr` to return a non-literal type.

This could be rectified by making `destroy()` return a dummy type that
is `constexpr`, however really there is no benefit in doing this
anyway. This has been fixed by simply removing `constexpr`, since any
context that wants to call `destroy()` will be unable to in a
`constexpr` way anyway.