✨ Add swap for `expected<T,E>`
The `E` value for `expected` is internalized within the abstraction.
This prevents direct access to the underlying reference, and thus also
prevents the ability to easily swap this value with another expected.

To make this possible, an overload of `swap` has been added for both
`expected<T,E>` and `expected<void,E>`