🔨 Remove deeply nested base classes
`expected` is constructed out of various base-classes that enable
conditional assignment and constructors. This composition is fine
on a technical level, however it does affect usability from the
debugger. When attempting to view the contained value of an `expected`,
it requires expanding through the ~9 base-classes to see the current
state of the value.

To make this less problematic for usability, these base classes
are now only conditionally added to the hiararchy if the given criteria
is triggered.