🩹 Fix support for expected<T&>
This implementation of `expected` was intended to support reference
types in addition to value types and void. This feature was added to
allow for APIs to return references in a possibly-failing manner without
requiring the heavy weight of exceptions, or the extra indirection of
pointers. Pointers, in particular, are not an equivalent exchange, since
it would ultimately result in the pointer having its own nullability
independent of the referenced value.

Since this was an intended feature, and it was not previously supported,
this support has been fixed.