🩹 Fix reference conversion construction/assign
`reference_wrapper<T>` is not officially constructible or
convert-assignable from `reference_wrapper<U>`. Although some
implementations *may* offer this as a functionality, this is not
standard C++ and may be subject to portability issues.

To attempt to fix this, the base class of `expected` has been
altered to conditionally call  `reference_wrapper::get()` if
`T` is a reference.