🔨 Rename inclusion guard
During part of the rework, the inclusion guard somehow came to be
renamed from `EXPECTED_EXPECTED_HPP` to be `MSL_EXPECTED_HPP`.

`MSL` is the name of another project I have been working on
independently, but it's unclear how this name snuck into the mix; must
have been a bad regex when switching editors. This fixes the guard to
be the correct name.