🎯 Add constexpr tests for expected
The base API of `expected` has various tests available for the general
runtime behavior, but has not yet been validated for the compile-time
behavior; whether it could be used in a `constexpr` context.

This adds some basic tests to ensure that a valid C++11 compiler will
be able to compile this as constant expressions.