🔧 Disable unit tests by default
Unit tests for this project are an optional developer-focused feature
that is not necessary for consumers of this project. For a consumer that
wishes to simply install this, or use it as a cmake subdirectory, it
is undesirable to force the unit tests on them.

To avoid forcing the unit tests on other developers, the flag that
enables them has been disabled by default, but may be enabled by
defining `EXPECTED_COMPILE_UNIT_TESTS=On` during cmake configuration.