🔨 Change base of bad_expected_access
The `bad_expected_access` was previously implementing `std::exception`
directly, which ultimately results in `what()` being implemented in this
header. By defining a virtual function in a header, this causes all
consumers to have the VTable emitted when building with GNU ABI -- which
is a nonzero cost.

It's undesirable to have a cost based solely on including the header.

To fix this, the base type has been changed from `std::exception` to
`std::logic_error`.