🔨 Update monadic functions to use rvalues
The original implementation, and intention, of the monadic API was for
only the *alternative* type to be passed as an rvalue, and for the
function mapping to perform the transformation into a new type.

However, after consideration, if the alternative type will already
undergo a form of remapping, it would make more sense for consistency
to *allow* the mapping functions to perform move-conversions as well
to allow for more optimal generation. This operation is already
destructive for the alternative, and thus may be destructive for the
mapped type as well for consistency.

As an added benefit, if the function performing the mapping chooses to
prefer const access, then the behavior will be the same as the previous
implementation.

In order to change all of these types, a new overload had to also be
added for the `void` specialization with `map_error`, which previously
did not have ref-qualified functions.