✨ Add flat_map_error
This adds an implementation of flat_map_error for use in the monadic
API.

This was initially omitted since, practically, this is unlikely to be
used in any functional way. However, due to the error type being
entirely encapsulated within `expected`'s API, there is no other way
to perform this form of mapping with rvalue-correctness unless
implemented as a function.

As a result, this has been added for completion; even though it's
unlikely to be practically used.