✨ Fully implement invoke_result
Previously `invoke_result` was only implemented in so far as basic
deduction without any form of SFINAE. This omission results in compile
errors if presented with an invocable overload that fails substituion.

As a result, this has been implemented fully such that the `::type` is
SFINAE disabled so that substitution failure does not cause a compile
error.