🩹 Fix gcc suboptimal code generation
GCC is generating suboptimal code generation during construction of
the `union` values that is resulting in several unnecessary
instructions being emitted. Returning a trivial `expected<int,int>`
with the value of `42` is resulting in the following instructions at
`-O1` or above:

```
mov     DWORD PTR [rsp-8], 42
mov     BYTE PTR [rsp-4], 1
mov     rax, QWORD PTR [rsp-8]
ret
```

Which can, and should be, optimized to:

```
movabs  rax, 4294967338
ret
```

Clang does not have this issue, and can optimize this case for us.

Inspection into GCC's behavior appears to indicate that the indirection
of forwarding arguments up a hierarchy is the cause of this generation.

If the `expected_destruct_base` is a base-class whose constructors are
invoked either through inheriting constructors _or_ through a custom
defined constructor that calls the parent class's constructor, then
this behavior is seen.

If the `expected_destruct_base` is a _member_ of a class that calls its
constructor instead, then we see the optimal code generation.

It's not clear what causes this discrepancy or why. Whatever the cause
may be, the fix appears to be to make the previous base-class exist as
a member of the derived class instead, which has been done here.