✏️ Rename `expected_destruct_base`
With the `expected_destruct_base` object being used in composition in
`expected_construct_base` now to fix the GCC code generation bug, the
name `*_base` does not truly represent its position anymore.

To address this, it has been renamed `expected_union` since this
effectively is a union around the `T` and `E` types.