📇 Rearrange badges
The badges for this repository have been rearranged, and some of them
have been dropped.

I don't really want the README to be crowded with unnecessary
information, and I'd rather make this as succinct as possible.
Since these details may be seen elsewhere, it's been decided that this
will simply be removed from the badge listing.