🧹 Move header inclusions to top of `expected.hpp`
As a general rule, I normally aim to keep all header inclusons as the
first thing at the top of a C++ header or source file. During
development at some point, the compile time definitions bubbled their
way to the top, which breaks this general style.

This relocates the headers to the top.