🔨 Remove default `E` argument
Previously, `expected`'s default `E` argument was `std::error_code`;
however, this decision is effectively arbitrary. In modern C++, there
are two competing types to represent errors; either `std::error_code`,
for representing system-specific error codes, or `std::error_condition`,
which represents system-agnostic error cases.

`std::error_code` is not inherently better or worse than
`std::error_condition` for this purpose. Internally, both of these
types are implemented as an `int` and a base-pointer to a
`std::error_category`. By defaulting `E` to *either* of these types, it
forces the size of the underlying type to be _at least_ large enough for
the `std::error_code` -- even if the underlying type is a fundamental
type that is smaller than this, which is undesirable.

Additional, although `error_code` is *convenient* for representing
error types, the type-erasure makes comparing an extracting the result
more ambiguous, and prevents APIs from semantically conveying the
set of error code in a clear way. It is clearer to return an explicit
enumeration type, or type which represents the error -- so that it may
be visible from the API.

As a result, rather than forcing `error_code` as the default, the
default itself has been removed, so that the user must specify it.
All tests have been updated to include this change as well.