🩹 Clean up unexpected conversion ctor/assign
The unexpected class's conversion constructor and assignment operators
have been updated to be more generic, using a template-deduced type with
forwarding rather than explicit underlying references. This allows us to
remove the SFINAE that was previously used to conditionally handle
reference types, so that this is only done exactly once.