✏️ Rename to shorter namespace
The project was previously defined under `namespace expect`, which
results in the rather unfortunate reading of certain types, such as:
`expect::unexpected`. This naming causes the unexpected type to itself
sound expected, which is counterintuitive. As a separate point, the
naming is more verbose than would be desirable.

In an effort to clean this up a little better, the namespace has been
changed from `expect` to simply` cpp`. Although `cpp` may be a
more common possible namespace, it should be *reasonably* safe to assume
that the name `expected` and `unexpected` should not collide. Inside of
this `namespace`, the definitions are under an
`inline namespace bitwizeshift` to help provide more uniqueness in the
(unlikely) case that these names ever result in a collision.