✏️ Rename `expected` to `result`
Among the C++ community, a monadic error-handling type is often
conventionally referred to as `expected<T>` -- something popularized
by (if I recall correctly) Andrei Alexandrescu's `Expected` type.
This has served as the baseline for several utilities of the same name,
along with the proposal for `std::expected` for which this entire
project was once based on.

However, in modern languages this utility goes by another, more
idiomatic name: `Result`. In other languages such as Rust, Swift, and
even Kotlin, error handling can be done through a monadic `Result` type
which effectively offers a similar/equivalent API as `expected` offers.