🧹 Undefine macros
The `expected.hpp` header has accumulated various macros that leak
out after inclusion, but are not an intended part of its API. Most of
these are symbols that only exist for the implementation of the
`expected` class.

To clean this up, these macros have all been `#undef`ed before
conclusion of the header.