✨ Attach error with bad_result_access
The exception thrown during access to a `result` object previously
simply contained a generic error message, but did not contain the
error code. This doesn't lend much context to the actual error, which
complicates identification by the user.

The `std::expected` proposal for which this project was once based on
throws an exception containing the `E` type, but derives from an
exception containing the `void` type. This specialization of the
exception is precisely what was being avoided by not doing this in the
first place as part of this implementation.

However, since designsh ave changed, it has become more apparent that it
would be useful to attach the exact error type into the exception for
the few cases that this will be used.