✨ Document `result` with nodiscard attribute
Since the `result` type encodes possible fallibility in APIs, its
important that the consumer actually be aware and _use_ this
information.

To ensure that this information is never unintentionally discarded,
the `result` class itself has been documented with the `[[nodiscard]]`
attribute when in C++17 mode, or the `[[clang::warn_unused_result]]`
attribute when using clang 3.9 or above. This ensures that any code
returning or using the `result` class will also, generate warnings when
ignored.

This attribute has been applied to both the class itself, as well as
constructors -- the latter of which ensures that no anonymous
constructions can occur without itself being inspected.

Unfortunately for GCC, this attribute only applies to function
return types, and is thus it cannot receive this diagnostic without
C++17 support.