✨ Warn on unused results from `result`
A modern C++ guideline is to document as much code as possible with the
`[[nodiscard]]` or equivalent attribute to ensure that no returned
results unintentionally get discarded or missed.

Since the `result` types whole purpose revolves around insepction of
the success/failure state, its undesirable for any of this data to
potentially be missed.

To ensure that this does not occur, the API has been documented with
the `[[nodiscard]]` attribute in C++17, or the
`[[clang::warn_unused_result]]` or `[[gnu::warn_unused_result]]`
attribute when in clang or GCC respectively.

This will work for all C++11-compatible GCC versions, and for clang
3.9 and above.

Unfortunately, MSVC does not appear to offer a similar requivalent
alternative outside of the SAL __checkReturn attribute, which
requires a specific compiler flag to operate correctly. MSVC will
require C++17 in order to support this feature.