🩹 Suppress "unused" warnings in unit tests
The unit tests on GCC are failing to compile due to the
`result::value()` expression being used in contexts where the result is
being ignored. This diagnostic is (effectively) correct, since the
context is checking either a `REQUIRE_THROWS` or `REQUIRED_NOTHROW`
expression, where we are only testing the behavior and not the value.

To fix this, any intentionally unused expressions have been wrapped
in a simple `suppress_unused` wrapper.