✨ Add `result::expect` for asserting value-state
With the recent addition of the `[[nodiscard]]` attribute and equivalent
on the `result` type, it has become more important to actually consume
a returned `result` value to avoid tripping this diagnostic.

One way to avoid this diagnostic is simply to store it to variable but
not use it, though this may not be the most straight-forward way to do
this.

The Rust programming language offers a functionality to take care of
this, called `expect` -- a function that basically acts as a simple
consumer of the error which will either proceed on success, or `panic!`
on failure. This approach can actually be modeled rather easily in
C++ just the same by providing two overloads, a `const &` and `&&`
qualified `expect` functions.

This has been added to the API so that a simple vocabulary can be
produced when a value is expected (asserted) to be correct, e.g.

```cpp
file::open("foo.txt").expect("Unable to open 'foo.txt'!");
```