🔧 Print to `stderr` if exceptions disabled
The current implementation of `result` will simply call `std::abort`
without any messages if `RESULT_DISABLE_EXCEPTIONS` is defined.

This provides no context for any errors to the caller, which can make
debugging difficult. Although logging is equally as undesirable,
tripping such a condition is considered a contract violation similar to
an assertion at this point, and the branch should never exist in
production.

To provide better clarity on failure, this will now print to `stderr`
on contract violation for both `value()` and `expect(...)` calls.