🩹 Fix C++11 `noexcept`-status of `swap(failure)`
The `swap` overload for `failure<E>` objects was marked as `noexcept`
when building this library with anything below C++17, due to the lack
of `std::is_nothrow_swappable`. This attribution is incorrect, and
may, in-fact, be harmful to incorrectly state.

To address this defect, the `noexcept` status of `swap` prior to C++17
will be deduced as `is_nothrow_move_constructible<E>::value`.