✨ Add support for `result<void,E>{in_place}`
The `result<void,E>` specialization has been notably lacking an
`in_place` overload due to the fact that `void` is not,
strictly-speaking, constructible.

However, this asymmetry is noticeable and unnecessary from the API, and
may result in diminished returns in terms of metaprogramming
composability.

In order to account for this, a new constructor of
`result<void,E>{in_place}` has been added which effectively
default-constructs the `result`.