🗑️ Remove constructor attributes
Using `[[nodiscard]]` on a constructor is apparently not as
well-supported as it had first seemed. Clang supports this case
easily, but GCC errors on it horribly.

Its not worth providing a functionality that exactly 1 compiler
supports, because this can massively complicate using the library
portably -- and so this has been removed entirely.