🩹 Correct `error_or` test
The `result<T,E>::error_or` test incorrectly contained a copy-paste
bug, in that it was actually testing the `result<void,E>`
implementation.

Since these two types are defined differently, with `void` being a
partial specialization, this is fundamentally testing a different
implementation -- which means ths previous implementation was
incorrectly not covering the test cases that it was supposed to be.

This fixes the issue.