🧹 Simplify `deploy-doxygen` workflow
The `deploy-doxygen` workflow was previously split into two
separate deployment scripts; one for release, and one for the "latest";
the latter of which only operates off of `master`.

This divide is mostly unnecessary, and in-fact is actually problematic
due to the code duplication, since it requires several places to be
updated any time fixes are made (which already resulted in an error
earlier during development, where the release script was not
updated to use `GITHUB_ENV`).

The two scripts have now been joined into one, which will instead
detect the trigger and determine the appropriate output path for
the generated documentation.