🔨 Disable erroneous warnings
`clang-cl`'s `-Wdocumentation` warnings, which are enabled by default
with `-Wall` due to peculiarities in the way `clang-cl` treats `-Wall`,
erroneously detects `\copydoc` tags as being invalid doxygen tags.

This diagnostic is *incorrect*, and has an open bug for it in the LLVM
bug tracker: https://bugs.llvm.org/show_bug.cgi?id=17437

Although this could be fixed by suppressing the warning from the
build system, this would directly affect any consumer that chooses to
use `clang` or `clang-cl` with this configuration -- which would
effectively be an API bug. Thus, the warnings have been addressed
through diagnostic pragmas in the header itself.