🩹 Address clang-cl's `-Wall` diagnostics
`clang-cl` is quite nit-picky with some of its warnings. Most of these
warnings are unimportant, but a few of them are legitimate.

All of these warnings apply solely to the unit-tests, and thus are not
strictly important. These have been addressed either directly through
minor code tweaks in the tests, or through disabling the warnings
in the build system (for implicit conversions, loss-of-data, etc).