🎯 Add missing tests for `failure`
The `failure` class' unit tests were rather rushed initially, which
missed many basic tests for constructors, assignment, and swapping.

This adds more tests for this type to ensure correctness.
In the process of some of these tests, errors have been discovered
and fixed in prior commits, such as improper SFINAE and the
recursive `swap` definition.