🩹 Fix `is_assignable` SFINAE conditions
Various SFINAE conditions that make use of `std::is_assignable` were
incorrectly using `T` as the assignable type, rather than a reference
`T&` type. This causes legitimate failures for substitutions due to
`T` being treated as a PR-value, which is not valid to assign to in most
cases.

This was not detected despite the comprehensive testing due to
the implicit construction and assignability from `result<T,E>`.

The various `is_assignable` constructions have been fixed.