🔨 Aggregate initialize `E` for `error()`
The `result<T,E>::error()` function returns a default-constructed `E`.
The intention for this library is that this would be aggregate
initialized to ensure that `E` is zero-initialized for primitive types.

This change already exists for `result<void,E>`, but was evidently
missed for `result<T,E>`. This has been updated to match the `void`
specialization.