🔨 Reduce number of generated matrix jobs
The workflows were modified recently to extract the `build_type` out
as a matrix input. This works well, however it generates twice as many
jobs -- many of which will not run concurrently on the default free
Github Actions runners.

This is resulting in longer execution times due in part to the
duplicate setup times. In reality, the debug/release build_type can
be better folded into the same action run, which reduces the number of
builds *by half* at the cost of minimal duplication of the config,
build, and test steps. This would allow the initial installation and
setup time to be saved, resulting only in the configuration/build/test
time.

This change has been made in an effort to reduce the CI times.