🗑 Remove conan deploy action
Bintray has been sunset, which means that the old conan repository there
is no longer valid to push to. The `conan-center` repository, which is
suggested as an alternative, works in a drastically different way as it
requires a custom conanfile defined for their manifests.

As a result, this whole workflow and any references have been removed