📇 Update compiler explorer links to refer to 1.0
The compiler explorer links in the README were previously all referring
to pinned SHA1s of the state of the repo at a given time. This replaces
all of these links with ones from the new `v1.0.0` release tag -- so
that these examples may be more illustrative of the most recent release.