Add dev CML file that loads all of the project
This file provides a containment for some gruesome details that don't
belong in other the project CML files, as these aren't descriptions of
the project, but rather wire things together properly to allow
convenient development of the library.

But that is okay, because that is exactly what superbuild CML files are
for, which this file is.