📇 Update commit standards to be less strict
The commit standards were previously written to strictly require all
contributors to use emoji-keys before contributing.

Although this is desirable for consistency, the reality is that many
people either won't bother to prefix the keys, or simply won't read this
requirement at all.

Due to this reality, I've lessened the strictness to only recommend
prefixing an emoji for categorization purposes, but to still require
the rest of the basic commit guidelines. The wording has also been
simplified and cleaned up, so that its not as intimidating.