✨ Add feature-preview banner
When this was released on Reddit, there was a comment that requested
more examples in the teaser of the repo to draw in more attention.
Unfortunately finding small enough examples to fit in a small code
snippet can be challenging; and so to accomplish this task, a new
`gif` image has instead been added that will become this repository's
banner -- replacing the teaser image.