🗑️ Remove `examples` and merge in with `tutorial`
The "examples" page was always left underwhelming with very little
real concrete examples of how this can be adequately used.
The majority of the real examples were always found under the
"tutorials" page, which includes tutorials for how to use the
entire API.

As a result, the examples from `example.md` have been relocated to
the `tutorials.md` page, and the examples.md has been removed.